Clean some warnings

+++ src/lib/time/ctime.c
@@ -5,6 +5,7 @@
  * @author Nikolay Korotky
  */
 #include <time.h>
+#include <stdio.h>
 
 #define EPOCH_START 1970
 #define YEAR_LENGHT 31556926
@@ -47,6 +48,6 @@ char *ctime(const time_t *t, char *buff) {
 		time -= MIN_LENGHT;
 		min++;
 	}
-	sprintf(buff, "%04d-%02d-%02d %02d:%02d:%02d", year, month, day, hour, min, time);
+	sprintf(buff, "%04d-%02d-%02d %02d:%02d:%02d", year, month, day, hour, min, (int)time);
 	return buff;
 }
